fix: 300行制限違反解決 - テストファイル分割とディレクトリ構造改善
## 主要修正

### 🔧 300行制限遵守
- **test_integration_coverage_boost.py**: 485行 → 削除・分割
- **test_performance_benchmarks.py**: 445行 → 削除・分割
- **最大ファイルサイズ**: 172行 (300行制限の57%)

### 📁 ディレクトリ構造改善
```
tests/
├── integration/           # 統合テスト (6ファイル)
│   ├── test_core_integration.py          (112行)
│   ├── test_config_command_integration.py (60行)
│   ├── test_gui_rendering_integration.py  (100行)
│   ├── test_debug_validation_integration.py (91行)
│   ├── test_utilities_endtoend_integration.py (161行)
│   └── test_template_integration.py       (30行)
└── performance/           # パフォーマンステスト (4ファイル)
    ├── test_basic_benchmarks.py           (172行)
    ├── test_memory_performance.py         (78行)
    ├── test_monitoring_benchmarks.py      (128行)
    └── test_realworld_scenarios.py        (112行)
```

### 🎯 TDDマーカー活用強化
- **マーカー使用率**: 142% (54マーカー/38テスト)
- **@pytest.mark.tdd_green**: Green Phase確認
- **@pytest.mark.tdd_refactor**: リファクタリング安全性確認
- **カテゴリ分類**: unit, integration, performance, slow, file_io等

### 🚀 機能保持
- **全テスト機能**: 統合・パフォーマンステスト完全保持
- **テストカバレッジ**: 17.10%維持
- **TDD基盤**: Red-Green-Refactorサイクル支援継続

### ✅ 品質確認
- **ファイル分割**: 機能別・責任分離
- **単一責任**: 各ファイル80-172行で明確な責任
- **コード重複なし**: 既存コード移動・整理のみ

---

🎯 **レビュー指摘対応完了**: 300行制限厳守 + 構造改善
