+++ tests/integration/__init__.py
@@ -1,0 +1,4 @@
+"""Integration Tests Package for Kumihan-Formatter
+
+統合テストパッケージ - モジュール間の連携とワークフロー全体をテスト
+"""
