+++ tests/performance/__init__.py
@@ -1,0 +1,4 @@
+"""Performance Tests Package for Kumihan-Formatter
+
+パフォーマンステストパッケージ - 処理速度、メモリ使用量、スループットの測定
+"""
