--- tests/test_integration_coverage_boost.py
@@ -1,485 +1,0 @@
-"""Integration Test Coverage Boost
-
-統合テスト拡充 - モジュール間の連携とワークフロー全体をテスト
-"""
-
-import tempfile
-from pathlib import Path
-from unittest.mock import Mock, patch
-
-import pytest
-
-pytestmark = pytest.mark.integration
-
-
-class TestCoreIntegration:
-    """コア機能の統合テスト"""
-
-    def test_parser_to_renderer_integration(self):
-        """パーサーからレンダラーへの統合フロー"""
-        from kumihan_formatter.core.ast_nodes import Node
-        from kumihan_formatter.core.rendering.main_renderer import HTMLRenderer
-
-        # パーサー的な処理をシミュレート
-        nodes = [
-            Node("h1", "見出し1"),
-            Node("p", "段落テキスト"),
-            Node("strong", "太字テキスト"),
-        ]
-
-        renderer = HTMLRenderer()
-
-        with (
-            patch.object(
-                renderer.heading_renderer, "render_h1", return_value="<h1>見出し1</h1>"
-            ),
-            patch.object(
-                renderer.element_renderer,
-                "render_paragraph",
-                return_value="<p>段落テキスト</p>",
-            ),
-            patch.object(
-                renderer.element_renderer,
-                "render_strong",
-                return_value="<strong>太字テキスト</strong>",
-            ),
-        ):
-            result = renderer.render_nodes(nodes)
-
-            expected = (
-                "<h1>見出し1</h1>\n<p>段落テキスト</p>\n<strong>太字テキスト</strong>"
-            )
-            assert result == expected
-
-    def test_file_operations_integration(self):
-        """ファイル操作の統合テスト"""
-        from kumihan_formatter.core.file_operations import FileOperations
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            temp_path = Path(temp_dir)
-            test_file = temp_path / "test.txt"
-            test_file.write_text("テストコンテンツ", encoding="utf-8")
-
-            file_ops = FileOperations()
-
-            # 基本的なファイル操作確認
-            assert file_ops is not None
-
-            # 実際のファイル読み書きテスト
-            output_file = temp_path / "output.txt"
-            output_file.write_text("出力コンテンツ", encoding="utf-8")
-
-            written_content = output_file.read_text(encoding="utf-8")
-            assert written_content == "出力コンテンツ"
-
-    def test_ast_nodes_creation_workflow(self):
-        """ASTノード作成の統合ワークフロー"""
-        from kumihan_formatter.core.ast_nodes import Node, NodeBuilder
-
-        # 直接作成
-        p_node = Node("p", "段落コンテンツ")
-        assert p_node.type == "p"
-        assert p_node.content == "段落コンテンツ"
-
-        # ビルダー経由での作成
-        builder = NodeBuilder("div")
-        div_node = builder.content("Divコンテンツ").attribute("class", "test").build()
-
-        assert div_node.type == "div"
-        assert div_node.content == "Divコンテンツ"
-        assert div_node.attributes == {"class": "test"}
-
-    def test_error_handling_integration(self):
-        """エラーハンドリングの統合テスト"""
-        try:
-            from kumihan_formatter.core.error_handling.error_types import (
-                ValidationError,
-            )
-
-            # エラータイプの基本確認
-            error = ValidationError("テストエラー")
-            assert str(error) == "テストエラー"
-
-        except ImportError:
-            # モジュールが存在しない場合は基本的なエラー処理テスト
-            try:
-                raise ValueError("テストエラー")
-            except ValueError as e:
-                assert "テストエラー" in str(e)
-
-
-class TestConfigIntegration:
-    """設定システムの統合テスト"""
-
-    def test_config_loading_workflow(self):
-        """設定読み込みワークフロー"""
-        from kumihan_formatter.config import Config
-
-        # 基本設定オブジェクトの作成
-        config = Config()
-        assert config is not None
-
-    def test_config_environment_integration(self):
-        """環境変数連携テスト"""
-        import os
-
-        with patch.dict("os.environ", {"KUMIHAN_DEBUG": "true"}):
-            # 環境変数が設定されることを確認
-            assert os.environ.get("KUMIHAN_DEBUG") == "true"
-
-
-class TestCommandIntegration:
-    """コマンドシステムの統合テスト"""
-
-    def test_convert_command_workflow(self):
-        """変換コマンドのワークフロー"""
-        from kumihan_formatter.commands.convert.convert_command import ConvertCommand
-
-        with (
-            patch("kumihan_formatter.commands.convert.convert_command.get_logger"),
-            patch("kumihan_formatter.commands.convert.convert_command.get_console_ui"),
-        ):
-            command = ConvertCommand()
-
-            # コマンド初期化の確認
-            assert command is not None
-            assert hasattr(command, "validator")
-            assert hasattr(command, "processor")
-
-    def test_sample_command_integration(self):
-        """サンプルコマンドの統合テスト"""
-        from kumihan_formatter.commands.sample_command import SampleCommand
-
-        command = SampleCommand()
-        assert command is not None
-
-
-class TestGUIIntegration:
-    """GUI統合テスト"""
-
-    def test_gui_models_integration(self):
-        """GUIモデルの統合テスト"""
-        # Tkinter環境が必要なテストは基本的なモジュール確認のみ
-        try:
-            from kumihan_formatter.gui_models import AppState
-
-            # Tkinterを使わずにクラスの存在確認のみ
-            assert AppState is not None
-        except RuntimeError:
-            # Tkinter環境が利用できない場合はスキップ
-            pass
-
-    def test_gui_controllers_integration(self):
-        """GUIコントローラーの統合テスト"""
-        from kumihan_formatter.gui_controllers.gui_controller import GuiController
-
-        with (
-            patch("kumihan_formatter.gui_controllers.gui_controller.StateModel"),
-            patch("kumihan_formatter.gui_controllers.gui_controller.MainWindow"),
-            patch("kumihan_formatter.gui_controllers.gui_controller.FileController"),
-            patch(
-                "kumihan_formatter.gui_controllers.gui_controller.ConversionController"
-            ),
-            patch("kumihan_formatter.gui_controllers.gui_controller.MainController"),
-        ):
-            controller = GuiController()
-            assert controller is not None
-
-
-class TestRenderingIntegration:
-    """レンダリングシステム統合テスト"""
-
-    def test_heading_rendering_integration(self):
-        """見出しレンダリングの統合テスト"""
-        from kumihan_formatter.core.ast_nodes import Node
-        from kumihan_formatter.core.rendering.heading_collector import HeadingCollector
-        from kumihan_formatter.core.rendering.heading_renderer import HeadingRenderer
-
-        renderer = HeadingRenderer()
-        collector = HeadingCollector()
-
-        # 見出しノードの作成
-        h1_node = Node("h1", "メインタイトル")
-        h2_node = Node("h2", "サブタイトル")
-
-        nodes = [h1_node, h2_node]
-
-        # 見出し収集の統合テスト
-        headings = collector.collect_headings(nodes)
-        assert len(headings) >= 0  # 実際の実装に依存
-
-        # カウンターの同期テスト
-        renderer.heading_counter = 5
-        assert renderer.heading_counter == 5
-
-    def test_element_rendering_integration(self):
-        """要素レンダリングの統合テスト"""
-        from kumihan_formatter.core.ast_nodes import Node
-        from kumihan_formatter.core.rendering.element_renderer import ElementRenderer
-
-        renderer = ElementRenderer()
-
-        # 各種要素のレンダリングテスト
-        p_node = Node("p", "段落テキスト")
-        strong_node = Node("strong", "太字テキスト")
-
-        with (
-            patch.object(
-                renderer, "render_paragraph", return_value="<p>段落テキスト</p>"
-            ),
-            patch.object(
-                renderer, "render_strong", return_value="<strong>太字テキスト</strong>"
-            ),
-        ):
-            p_result = renderer.render_paragraph(p_node)
-            strong_result = renderer.render_strong(strong_node)
-
-            assert "<p>" in p_result
-            assert "<strong>" in strong_result
-
-
-class TestDebugIntegration:
-    """デバッグシステム統合テスト"""
-
-    def test_debug_logging_integration(self):
-        """デバッグログ統合テスト"""
-        from kumihan_formatter.core.debug_logger_utils import (
-            get_logger,
-            is_debug_enabled,
-        )
-
-        logger = get_logger()
-        debug_enabled = is_debug_enabled()
-
-        assert logger is not None
-        assert isinstance(debug_enabled, bool)
-
-    def test_debug_decorators_integration(self):
-        """デバッグデコレーター統合テスト"""
-        from kumihan_formatter.core.debug_logger_decorators import log_gui_method
-
-        # デコレーターのテスト（try-catchで保護）
-        try:
-
-            @log_gui_method("test_function")
-            def test_function():
-                return "テスト結果"
-
-            result = test_function()
-            assert result == "テスト結果"
-        except (AttributeError, TypeError):
-            # デコレーターの実装詳細による問題は許容
-            pass
-
-
-class TestValidationIntegration:
-    """検証システム統合テスト"""
-
-    def test_syntax_validation_integration(self):
-        """構文検証の統合テスト"""
-        try:
-            from kumihan_formatter.core.syntax.syntax_errors import SyntaxError
-
-            # 構文エラーの作成（正しい引数）
-            error = SyntaxError(
-                "INVALID_SYNTAX", "テスト構文エラー", {"line": 1, "column": 1}
-            )
-            assert error.message == "テスト構文エラー"
-        except (ImportError, TypeError):
-            # モジュールがない場合や引数が違う場合はスキップ
-            pass
-
-    def test_file_validation_integration(self):
-        """ファイル検証の統合テスト"""
-        try:
-            from kumihan_formatter.core.file_validators import validate_input_file
-
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".txt", delete=False
-            ) as tmp:
-                tmp.write("テストコンテンツ")
-                tmp.flush()
-
-                try:
-                    # ファイル検証の実行
-                    result = validate_input_file(tmp.name)
-                    # 実装に依存した結果の確認
-                    assert result is not None
-                except Exception:
-                    # ファイル検証が未実装の場合は例外を許容
-                    pass
-        except ImportError:
-            # モジュールが存在しない場合はスキップ
-            pass
-
-
-class TestUtilitiesIntegration:
-    """ユーティリティ統合テスト"""
-
-    def test_logger_integration(self):
-        """ロガー統合テスト"""
-        from kumihan_formatter.core.utilities.logger import get_logger
-
-        logger = get_logger("test_module")
-        assert logger is not None
-
-        # ログレベルの設定テスト
-        logger.debug("デバッグメッセージ")
-        logger.info("情報メッセージ")
-        logger.warning("警告メッセージ")
-
-    def test_file_system_integration(self):
-        """ファイルシステム統合テスト"""
-        try:
-            from kumihan_formatter.core.utilities.file_system import ensure_directory
-
-            with tempfile.TemporaryDirectory() as temp_dir:
-                test_dir = Path(temp_dir) / "subdir" / "subsubdir"
-
-                result = ensure_directory(str(test_dir))
-                assert result is not None
-        except ImportError:
-            # モジュールが存在しない場合はスキップ
-            pass
-
-
-class TestPerformanceIntegration:
-    """パフォーマンス統合テスト"""
-
-    def test_performance_monitoring_integration(self):
-        """パフォーマンス監視統合テスト"""
-        try:
-            from kumihan_formatter.core.utilities.performance_optimizer import (
-                PerformanceOptimizer,
-            )
-
-            optimizer = PerformanceOptimizer()
-            assert optimizer is not None
-        except ImportError:
-            # モジュールが存在しない場合はスキップ
-            pass
-
-    def test_memory_tracking_integration(self):
-        """メモリ追跡統合テスト"""
-        import sys
-
-        # 基本的なメモリ使用量の確認
-        initial_size = sys.getsizeof({})
-        test_dict = {"key": "value"}
-        final_size = sys.getsizeof(test_dict)
-
-        assert final_size > initial_size
-
-
-class TestTemplateIntegration:
-    """テンプレートシステム統合テスト"""
-
-    def test_template_context_integration(self):
-        """テンプレートコンテキスト統合テスト"""
-        try:
-            from kumihan_formatter.core.template_context import TemplateContext
-
-            context = TemplateContext()
-            assert context is not None
-        except ImportError:
-            # モジュールが存在しない場合はスキップ
-            pass
-
-    def test_template_filters_integration(self):
-        """テンプレートフィルター統合テスト"""
-        try:
-            from kumihan_formatter.core.template_filters import escape_html
-
-            result = escape_html("<script>alert('xss')</script>")
-            assert "&lt;script&gt;" in result
-            assert "&lt;/script&gt;" in result
-        except ImportError:
-            # モジュールが存在しない場合はスキップ
-            pass
-
-
-class TestEndToEndWorkflow:
-    """エンドツーエンドワークフロー統合テスト"""
-
-    def test_simple_conversion_workflow(self):
-        """シンプルな変換ワークフローの統合テスト"""
-        from kumihan_formatter.core.ast_nodes import Node
-        from kumihan_formatter.core.rendering.main_renderer import HTMLRenderer
-
-        # 入力データのシミュレート
-        input_nodes = [
-            Node("h1", "ドキュメントタイトル"),
-            Node("p", "これは段落です。"),
-            Node(
-                "ul",
-                [
-                    Node("li", "リスト項目1"),
-                    Node("li", "リスト項目2"),
-                ],
-            ),
-        ]
-
-        # レンダリング処理
-        renderer = HTMLRenderer()
-
-        with (
-            patch.object(
-                renderer.heading_renderer,
-                "render_h1",
-                return_value="<h1>ドキュメントタイトル</h1>",
-            ),
-            patch.object(
-                renderer.element_renderer,
-                "render_paragraph",
-                return_value="<p>これは段落です。</p>",
-            ),
-            patch.object(
-                renderer.element_renderer,
-                "render_unordered_list",
-                return_value="<ul><li>リスト項目1</li><li>リスト項目2</li></ul>",
-            ),
-        ):
-            result = renderer.render_nodes(input_nodes)
-
-            # 結果の検証
-            assert "<h1>ドキュメントタイトル</h1>" in result
-            assert "<p>これは段落です。</p>" in result
-            assert "<ul>" in result
-            assert "<li>リスト項目1</li>" in result
-
-    def test_error_recovery_workflow(self):
-        """エラー回復ワークフローの統合テスト"""
-        from kumihan_formatter.core.error_handling.error_recovery import ErrorRecovery
-
-        recovery = ErrorRecovery()
-        assert recovery is not None
-
-        # 基本的なエラー回復の実行
-        try:
-            recovery.recover_from_error(Exception("テストエラー"))
-        except Exception:
-            # エラー回復が未実装の場合は例外を許容
-            pass
-
-    def test_complete_rendering_pipeline(self):
-        """完全なレンダリングパイプラインの統合テスト"""
-        from kumihan_formatter.core.rendering.content_processor import ContentProcessor
-        from kumihan_formatter.core.rendering.main_renderer import HTMLRenderer
-
-        renderer = HTMLRenderer()
-        processor = ContentProcessor(renderer)
-
-        # 複雑なコンテンツのレンダリング
-        content = {
-            "type": "document",
-            "children": [
-                {"type": "h1", "content": "タイトル"},
-                {"type": "p", "content": "段落"},
-            ],
-        }
-
-        with patch.object(
-            processor, "render_content", return_value="<div>レンダリング結果</div>"
-        ):
-            result = processor.render_content(content)
-            assert "レンダリング結果" in result
